backend/packages/wps-tools/pyproject.toml
[project]
name = "wps-tools"
version = "0.1.0"
description = "Wildfire Predictive Services Unit Tools"
authors = [{ name = "Conor Brady", email = "conor.brady@gov.bc.ca" }]
requires-python = ">=3.12.3,<4.0"
dependencies = [
    "python-decouple>=3.6,<4",
    "aiobotocore>=2.4.1,<3",
    "types-aiobotocore[s3]>=2.4.1,<3",
    "pandas>=1.5.2,<3",
    "requests>=2.31.0,<3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wps_tools"]
